flocculation: only use p-value cutoffs
According to the limma reference manual, including both a p-value and an
log-fold-change cutoff in topTable can increase the FDR above the
nominal level if they (p-values & lfcs) are not highly correlated.
According to the same source, the same goes for decideTests; the FDR or
family-wise error rate may also rise above the nominal level.